+++ config-loader/src/main/java/org/edmcouncil/spec/fibo/config/configuration/model/AppConfiguration.java
@@ -1,5 +1,6 @@
 package org.edmcouncil.spec.fibo.config.configuration.model;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
 import javax.annotation.PostConstruct;
@@ -22,8 +23,10 @@ public class AppConfiguration {
   @Autowired
   private FileSystemManager fileSystemManager;
   private ViewerCoreConfiguration viewerCoreConfig;
+  private final FileSystemManager fsm;
 
-  public AppConfiguration() {
+  public AppConfiguration(FileSystemManager fsm) {
+    this.fsm = fsm;
   }
 
   @PostConstruct
@@ -34,13 +37,27 @@ public class AppConfiguration {
     Path configFilePath = null;
 
     try {
-      configFilePath = fileSystemManager.getPathToWeaselConfigFile();
+
+      configFilePath = fileSystemManager.getPathToConfigFile();
+
+      LOG.debug("Path to Configs Directory : {}", configFilePath.toAbsolutePath().toString());
+      LOG.debug("Load config");
+      LOG.debug("List Files : {}", configFilePath.toFile().listFiles().toString());
+      for (File file : configFilePath.toFile().listFiles()) {
+        LOG.debug("Path to ConfigFile : {}", file.toPath().toString());
+        if (file.isFile()) {
+
+          cl.loadWeaselConfiguration(file.toPath());
+        }
+      }
+
+      this.viewerCoreConfig = cl.getConfiguration();
 
     } catch (IOException ex) {
       LOG.error("[ERROR] IOException while loading config file");
     }
 
-    this.viewerCoreConfig = cl.loadWeaselConfiguration(configFilePath);
+    this.viewerCoreConfig = cl.getConfiguration();
 
     if (!viewerCoreConfig.isEmpty()) {
       LOG.debug("Configuration: ");