+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+**## Unreleased [0.3.0] 
+**### Changed**
+- rebuilding the configuration file and its loader
+- splitting the configuration file into 5 files
+
 
 ## Unreleased [0.3.0]
 ### Added